requirements.txt
openai==1.3.8
gradio==3.50.2
langchain==0.0.349
pypdf==3.17.2
chromadb==0.4.18
tiktoken==0.5.2
weaviate-client==3.25.3
boto3==1.34.2
unstructured==0.11.6
playwright==1.40.0